+++ spec/rubocop/cli/cli_autocorrect_spec.rb
@@ -807,14 +807,12 @@ describe RuboCop::CLI, :isolated_environment do
 
   it 'does not say [Corrected] if correction was avoided' do
     src = ['# encoding: utf-8',
-           'not a && b',
            'func a do b end',
            "Signal.trap('TERM') { system(cmd); exit }",
            'def self.some_method(foo, bar: 1)',
            '  log.debug(foo)',
            'end']
     corrected = ['# encoding: utf-8',
-                 'not a && b',
                  'func a do b end',
                  "Signal.trap('TERM') { system(cmd); exit }",
                  'def self.some_method(foo, bar: 1)',
@@ -822,11 +820,10 @@ describe RuboCop::CLI, :isolated_environment do
                  'end']
     offenses =
       ['== example.rb ==',
-       'C:  2:  1: Use ! instead of not.',
-       'C:  3:  8: Prefer {...} over do...end for single-line blocks.',
-       'C:  4: 34: Do not use semicolons to terminate expressions.',
-       'W:  5: 27: Unused method argument - bar.']
-    summary = '1 file inspected, 4 offenses detected'
+       'C:  2:  8: Prefer {...} over do...end for single-line blocks.',
+       'C:  3: 34: Do not use semicolons to terminate expressions.',
+       'W:  4: 27: Unused method argument - bar.']
+    summary = '1 file inspected, 3 offenses detected'
     create_file('.rubocop.yml', ['AllCops:',
                                  '  TargetRubyVersion: 2.0'])
     create_file('example.rb', src)