+++ spec/rubocop/cop/style/not_spec.rb
@@ -26,9 +26,33 @@ describe RuboCop::Cop::Style::Not, :config do
     expect(new_source).to eq('!(test)')
   end
 
-  it 'leaves "not" as is if auto-correction changes the meaning' do
+  it 'uses the reverse operator when `not` is applied to a comparison' do
     src = 'not x < y'
     new_source = autocorrect_source(cop, src)
-    expect(new_source).to eq(src)
+    expect(new_source).to eq('x >= y')
+  end
+
+  it 'parenthesizes when `not` would change the meaning of a binary exp' do
+    src = 'not a >> b'
+    new_source = autocorrect_source(cop, src)
+    expect(new_source).to eq('!(a >> b)')
+  end
+
+  it 'parenthesizes when `not` is applied to a ternary op' do
+    src = 'not a ? b : c'
+    new_source = autocorrect_source(cop, src)
+    expect(new_source).to eq('!(a ? b : c)')
+  end
+
+  it 'parenthesizes when `not` is applied to and' do
+    src = 'not a && b'
+    new_source = autocorrect_source(cop, src)
+    expect(new_source).to eq('!(a && b)')
+  end
+
+  it 'parenthesizes when `not` is applied to or' do
+    src = 'not a || b'
+    new_source = autocorrect_source(cop, src)
+    expect(new_source).to eq('!(a || b)')
   end
 end